refactor(lsp): split indexing tests
