+++ crates/lsp/src/tests/mod.rs
@@ -7,19 +7,18 @@ use crate::{
 };
 use async_lsp::{ClientSocket, ErrorCode, ResponseError, ServerSocket, router::Router};
 use lsp_types::{
-    CodeLensWorkspaceClientCapabilities, CreateFilesParams, DeleteFilesParams, Diagnostic,
-    DiagnosticWorkspaceClientCapabilities, DidChangeConfigurationParams,
-    DidChangeTextDocumentParams, DidChangeWatchedFilesClientCapabilities,
-    DidChangeWatchedFilesParams, DidChangeWorkspaceFoldersParams, DidCloseTextDocumentParams,
-    DidOpenTextDocumentParams, DidSaveTextDocumentParams, DocumentDiagnosticParams,
-    DocumentDiagnosticReport, DocumentDiagnosticReportResult, DocumentSymbol, FileChangeType,
-    FileCreate, FileDelete, FileEvent, InlayHintWorkspaceClientCapabilities, PartialResultParams,
-    Position, ProgressParams, ProgressParamsValue, PublishDiagnosticsParams, Range,
-    RegistrationParams, SymbolKind, TextDocumentContentChangeEvent, TextDocumentIdentifier,
-    TextDocumentItem, UnregistrationParams, VersionedTextDocumentIdentifier, WatchKind,
-    WorkDoneProgress, WorkDoneProgressCreateParams, WorkDoneProgressParams,
-    WorkspaceClientCapabilities, WorkspaceFolder, WorkspaceFoldersChangeEvent, WorkspaceSymbol,
-    notification, notification::Notification, request,
+    CodeLensWorkspaceClientCapabilities, Diagnostic, DiagnosticWorkspaceClientCapabilities,
+    DidChangeConfigurationParams, DidChangeTextDocumentParams, DidChangeWatchedFilesParams,
+    DidChangeWorkspaceFoldersParams, DidCloseTextDocumentParams, DidOpenTextDocumentParams,
+    DidSaveTextDocumentParams, DocumentDiagnosticParams, DocumentDiagnosticReport,
+    DocumentDiagnosticReportResult, DocumentSymbol, FileChangeType, FileEvent,
+    InlayHintWorkspaceClientCapabilities, PartialResultParams, Position, ProgressParams,
+    ProgressParamsValue, PublishDiagnosticsParams, Range, SymbolKind,
+    TextDocumentContentChangeEvent, TextDocumentIdentifier, TextDocumentItem,
+    VersionedTextDocumentIdentifier, WorkDoneProgress, WorkDoneProgressCreateParams,
+    WorkDoneProgressParams, WorkspaceClientCapabilities, WorkspaceFolder,
+    WorkspaceFoldersChangeEvent, WorkspaceSymbol, notification, notification::Notification,
+    request,
 };
 use std::{
     future::Future,
@@ -43,6 +42,7 @@ mod folding_range;
 mod goto_definition;
 mod hover;
 mod implementation;
+mod indexing;
 mod inlay_hint;
 #[path = "protocol_trace.rs"]
 mod protocol_trace_tests;
@@ -51,9 +51,11 @@ mod refresh;
 mod rename;
 mod selection_range;
 mod signature_help;
+mod source_events;
 mod support;
 mod type_definition;
 mod type_hierarchy;
+mod watched_files;
 mod workspace_diagnostic;
 
 const ASYNC_TEST_TIMEOUT: Duration = Duration::from_secs(5);
@@ -63,81 +65,6 @@ enum WorkDoneEvent {
     Create(WorkDoneProgressCreateParams),
     Progress(ProgressParams),
     Diagnostics(PublishDiagnosticsParams),
-}
-
-#[derive(Debug)]
-enum WatchedFileClientEvent {
-    Register(RegistrationParams),
-    Unregister(UnregistrationParams),
-}
-
-async fn next_watched_file_client_event(
-    events: &mut mpsc::UnboundedReceiver<WatchedFileClientEvent>,
-) -> WatchedFileClientEvent {
-    tokio::time::timeout(ASYNC_TEST_TIMEOUT, events.recv())
-        .await
-        .expect("watched-file client event should arrive")
-        .expect("watched-file client event channel should stay open")
-}
-
-fn watched_file_registration_has_spec(
-    params: &RegistrationParams,
-    base: &Path,
-    pattern: &str,
-) -> bool {
-    let base_uri = Url::from_file_path(base).unwrap().to_string();
-    params.registrations.iter().any(|registration| {
-        registration.register_options.as_ref().is_some_and(|options| {
-            options["watchers"].as_array().is_some_and(|watchers| {
-                watchers.iter().any(|watcher| {
-                    watcher["globPattern"]["baseUri"].as_str() == Some(&base_uri)
-                        && watcher["globPattern"]["pattern"].as_str() == Some(pattern)
-                })
-            })
-        })
-    })
-}
-
-fn watched_file_registration_spec_kind(
-    params: &RegistrationParams,
-    base: &Path,
-    pattern: &str,
-) -> Option<u64> {
-    let base_uri = Url::from_file_path(base).unwrap().to_string();
-    params.registrations.iter().find_map(|registration| {
-        registration.register_options.as_ref().and_then(|options| {
-            options["watchers"].as_array().and_then(|watchers| {
-                watchers.iter().find_map(|watcher| {
-                    (watcher["globPattern"]["baseUri"].as_str() == Some(&base_uri)
-                        && watcher["globPattern"]["pattern"].as_str() == Some(pattern))
-                    .then(|| watcher["kind"].as_u64())
-                    .flatten()
-                })
-            })
-        })
-    })
-}
-
-fn watched_file_registration_has_recursive_spec_covering(
-    params: &RegistrationParams,
-    path: &Path,
-) -> bool {
-    params.registrations.iter().any(|registration| {
-        registration.register_options.as_ref().is_some_and(|options| {
-            options["watchers"].as_array().is_some_and(|watchers| {
-                watchers.iter().any(|watcher| {
-                    matches!(
-                        watcher["globPattern"]["pattern"].as_str(),
-                        Some("**/*.sol" | "**/foundry.toml")
-                    ) && watcher["globPattern"]["baseUri"]
-                        .as_str()
-                        .and_then(|uri| Url::parse(uri).ok())
-                        .and_then(|uri| uri.to_file_path().ok())
-                        .is_some_and(|base| path.starts_with(base))
-                })
-            })
-        })
-    })
 }
 
 struct WorkDoneClientState {
@@ -369,685 +296,6 @@ fn analysis_collects_open_versions_and_loaded_dependencies() {
         result.analyzed_documents,
         AnalyzedDocuments::from_iter([(main_uri, Some(0)), (dependency_uri, None)])
     );
-}
-
-#[test]
-fn analysis_tracks_excluded_transitive_dependencies_and_normalized_missing_candidates() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        import "./generated/B.sol";
-        contract Main is B {}
-
-        //- /generated/B.sol
-        import "./nested/../C.sol";
-        contract B is C {}
-
-        //- /generated/C.sol
-        import "./missing/../Missing.sol";
-        contract C {}
-
-        //- /generated/Unrelated.sol
-        contract Unrelated {}
-        "#,
-    );
-    let config = config_with_indexing_excludes(&project, &["generated/**"]);
-    let mut batches = snapshot_with_config(config, project.vfs()).analysis_batches(Vec::new());
-    let batch = batches.pop().unwrap();
-    assert!(batches.is_empty());
-    assert_eq!(batch.seen_paths, FxHashSet::from_iter([project.path("/Main.sol")]));
-
-    let output = analyze_cancellable(batch, &IndexingCancellation::default()).unwrap();
-
-    assert_eq!(
-        output.analysis_paths.resolved_dependencies,
-        FxHashSet::from_iter([project.path("/generated/B.sol"), project.path("/generated/C.sol")])
-    );
-    assert_eq!(
-        output.analysis_paths.missing_candidates,
-        FxHashSet::from_iter([project.path("/generated/Missing.sol")])
-    );
-    assert!(output.analysis_paths.existing_unresolved_candidates.is_empty());
-    assert!(
-        output.result.symbol_tables.workspace_symbols("B").iter().any(|symbol| symbol.name == "B")
-    );
-    assert!(
-        output.result.symbol_tables.workspace_symbols("C").iter().any(|symbol| symbol.name == "C")
-    );
-    assert!(output.result.symbol_tables.workspace_symbols("Unrelated").is_empty());
-}
-
-#[test]
-fn analysis_output_accumulator_resolved_path_wins_across_batches() {
-    let path = PathBuf::from("Dependency.sol");
-    let result = || AnalysisResult {
-        analyzed_documents: AnalyzedDocuments::default(),
-        diagnostics: DiagnosticMap::default(),
-        symbol_tables: SymbolTables::default(),
-    };
-    let mut accumulator = AnalysisOutputAccumulator::default();
-    accumulator.push(AnalysisOutput {
-        result: result(),
-        analysis_paths: AnalysisPathIndex {
-            missing_candidates: FxHashSet::from_iter([path.clone()]),
-            ..Default::default()
-        },
-    });
-    accumulator.push(AnalysisOutput {
-        result: result(),
-        analysis_paths: AnalysisPathIndex {
-            resolved_dependencies: FxHashSet::from_iter([path.clone()]),
-            ..Default::default()
-        },
-    });
-
-    let output = accumulator.finish();
-
-    assert_eq!(output.analysis_paths.resolved_dependencies, FxHashSet::from_iter([path]));
-    assert!(output.analysis_paths.existing_unresolved_candidates.is_empty());
-    assert!(output.analysis_paths.missing_candidates.is_empty());
-}
-
-#[test]
-fn stale_analysis_does_not_replace_published_path_index() {
-    let current_path = PathBuf::from("Current.sol");
-    let stale_path = PathBuf::from("Stale.sol");
-    let output = |path| AnalysisOutput {
-        result: AnalysisResult {
-            analyzed_documents: AnalyzedDocuments::default(),
-            diagnostics: DiagnosticMap::default(),
-            symbol_tables: SymbolTables::default(),
-        },
-        analysis_paths: AnalysisPathIndex {
-            resolved_dependencies: FxHashSet::from_iter([path]),
-            ..Default::default()
-        },
-    };
-    let state = GlobalState::new(ClientSocket::new_closed());
-    assert!(state.snapshot().publish_analysis_output(0, output(current_path.clone())));
-    let mut stale_snapshot = state.snapshot();
-    state.mark_analysis_pending_for_test();
-
-    assert!(!stale_snapshot.publish_analysis_output(0, output(stale_path)));
-    assert_eq!(
-        state.analysis_commit.lock().analysis_paths.resolved_dependencies,
-        FxHashSet::from_iter([current_path])
-    );
-}
-
-#[test]
-fn deferred_dependency_change_prevents_stale_analysis_publish() {
-    let state = GlobalState::new(ClientSocket::new_closed());
-    state.mark_analysis_pending_for_test();
-    let version = state.analysis_version.load(Ordering::Acquire);
-    let path = PathBuf::from("Dependency.sol");
-    assert_eq!(
-        state.classify_source_file_event(&path, FileChangeType::CHANGED),
-        SourceFileEventDisposition::Deferred
-    );
-    let output = AnalysisOutput {
-        result: AnalysisResult {
-            analyzed_documents: AnalyzedDocuments::default(),
-            diagnostics: DiagnosticMap::default(),
-            symbol_tables: SymbolTables::default(),
-        },
-        analysis_paths: AnalysisPathIndex {
-            resolved_dependencies: FxHashSet::from_iter([path]),
-            ..Default::default()
-        },
-    };
-
-    assert!(!state.snapshot().publish_analysis_output(version, output));
-    assert_eq!(*state.published_analysis_version.borrow(), 0);
-    assert!(state.analysis_commit.lock().deferred_source_file_events.is_empty());
-}
-
-#[test]
-fn source_event_classification_observes_path_index_committed_before_its_lock() {
-    let state = GlobalState::new(ClientSocket::new_closed());
-    state.mark_analysis_pending_for_test();
-    let version = state.analysis_version.load(Ordering::Acquire);
-    let path = PathBuf::from("Dependency.sol");
-    let mut commit = state.analysis_commit.lock();
-
-    std::thread::scope(|scope| {
-        let (started_tx, started_rx) = std_mpsc::sync_channel(1);
-        let (result_tx, result_rx) = std_mpsc::sync_channel(1);
-        let event_state = &state;
-        let event_path = path.clone();
-        scope.spawn(move || {
-            started_tx.send(()).unwrap();
-            result_tx
-                .send(event_state.classify_source_file_event(&event_path, FileChangeType::CHANGED))
-                .unwrap();
-        });
-        started_rx.recv_timeout(ASYNC_TEST_TIMEOUT).unwrap();
-
-        commit.analysis_paths.resolved_dependencies.insert(path.clone());
-        state.published_analysis_version.send_replace(version);
-        drop(commit);
-
-        assert_eq!(
-            result_rx.recv_timeout(ASYNC_TEST_TIMEOUT).unwrap(),
-            SourceFileEventDisposition::Relevant
-        );
-    });
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_file_specs_are_committed_before_the_next_analysis_epoch() {
-    let project = TestProject::new();
-    std::fs::create_dir(project.path("/workspace")).unwrap();
-    let mut params = project.initialize_params_with_roots(&["/workspace"]);
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, config) = negotiate_capabilities(params);
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    state.mark_analysis_pending_for_test();
-    let version = state.analysis_version.load(Ordering::Acquire);
-    let mut snapshot = state.snapshot();
-    let output = AnalysisOutput {
-        result: AnalysisResult {
-            analyzed_documents: AnalyzedDocuments::default(),
-            diagnostics: DiagnosticMap::default(),
-            symbol_tables: SymbolTables::default(),
-        },
-        analysis_paths: AnalysisPathIndex {
-            resolved_dependencies: FxHashSet::from_iter([project.path("/outside/Dependency.sol")]),
-            ..Default::default()
-        },
-    };
-    let desired_specs = state.watched_file_registration.desired_specs.lock();
-    let runtime = tokio::runtime::Handle::current();
-
-    std::thread::scope(|scope| {
-        let publisher = scope.spawn(move || {
-            let _runtime = runtime.enter();
-            snapshot.publish_analysis_output(version, output)
-        });
-
-        let deadline = Instant::now() + ASYNC_TEST_TIMEOUT;
-        while *state.published_analysis_version.borrow() != version && Instant::now() < deadline {
-            std::thread::yield_now();
-        }
-        assert_eq!(*state.published_analysis_version.borrow(), version);
-
-        let deadline = Instant::now() + Duration::from_millis(100);
-        let mut commit_became_available = false;
-        while Instant::now() < deadline {
-            if state.analysis_commit.try_lock().is_some() {
-                commit_became_available = true;
-                break;
-            }
-            std::thread::yield_now();
-        }
-
-        drop(desired_specs);
-        assert!(publisher.join().unwrap());
-        assert!(
-            !commit_became_available,
-            "analysis commit unlocked before its watched-file specs were queued"
-        );
-    });
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn workspace_folder_change_advances_epoch_before_watcher_reregistration() {
-    let project = TestProject::new();
-    let old_root = project.path("/old");
-    let new_root = project.path("/new");
-    std::fs::create_dir(&old_root).unwrap();
-    std::fs::create_dir(&new_root).unwrap();
-    let mut params = project.initialize_params_with_roots(&["/old"]);
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, config) = negotiate_capabilities(params);
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    let initial_version = state.analysis_version.load(Ordering::Acquire);
-    let registration = state.watched_file_registration.clone();
-    let desired_specs = registration.desired_specs.lock();
-    let analysis_version = state.analysis_version.clone();
-    let runtime = tokio::runtime::Handle::current();
-    let worker = std::thread::spawn(move || {
-        let _runtime = runtime.enter();
-        let result = crate::handlers::did_change_workspace_folders(
-            &mut state,
-            DidChangeWorkspaceFoldersParams {
-                event: WorkspaceFoldersChangeEvent {
-                    added: vec![WorkspaceFolder {
-                        uri: Url::from_file_path(new_root).unwrap(),
-                        name: "new".into(),
-                    }],
-                    removed: vec![WorkspaceFolder {
-                        uri: Url::from_file_path(old_root).unwrap(),
-                        name: "old".into(),
-                    }],
-                },
-            },
-        );
-        assert!(matches!(result, ControlFlow::Continue(())));
-        state
-    });
-
-    let deadline = Instant::now() + Duration::from_millis(100);
-    while analysis_version.load(Ordering::Acquire) == initial_version && Instant::now() < deadline {
-        std::thread::yield_now();
-    }
-    let advanced_before_reregistration =
-        analysis_version.load(Ordering::Acquire) != initial_version;
-
-    drop(desired_specs);
-    let state = worker.join().unwrap();
-    state.analysis_scheduler.tasks.lock().cancel();
-    assert!(
-        advanced_before_reregistration,
-        "workspace-folder change queued watchers before invalidating the old analysis epoch"
-    );
-}
-
-#[test]
-fn deferred_existing_missing_candidate_change_prevents_stale_analysis_publish() {
-    let project = TestProject::new();
-    let path = project.path("/Missing.sol");
-    project.write_file("/Missing.sol", "contract Missing {}");
-    let state = GlobalState::new(ClientSocket::new_closed());
-    state.mark_analysis_pending_for_test();
-    let version = state.analysis_version.load(Ordering::Acquire);
-    assert_eq!(
-        state.classify_source_file_event(&path, FileChangeType::CHANGED),
-        SourceFileEventDisposition::Deferred
-    );
-    let output = AnalysisOutput {
-        result: AnalysisResult {
-            analyzed_documents: AnalyzedDocuments::default(),
-            diagnostics: DiagnosticMap::default(),
-            symbol_tables: SymbolTables::default(),
-        },
-        analysis_paths: AnalysisPathIndex {
-            missing_candidates: FxHashSet::from_iter([path]),
-            ..Default::default()
-        },
-    };
-
-    assert!(!state.snapshot().publish_analysis_output(version, output));
-    assert_eq!(*state.published_analysis_version.borrow(), 0);
-}
-
-#[test]
-fn deferred_unrelated_change_does_not_block_analysis_publish() {
-    let state = GlobalState::new(ClientSocket::new_closed());
-    state.mark_analysis_pending_for_test();
-    let version = state.analysis_version.load(Ordering::Acquire);
-    assert_eq!(
-        state.classify_source_file_event(Path::new("Unrelated.sol"), FileChangeType::CHANGED),
-        SourceFileEventDisposition::Deferred
-    );
-    let output = AnalysisOutput {
-        result: AnalysisResult {
-            analyzed_documents: AnalyzedDocuments::default(),
-            diagnostics: DiagnosticMap::default(),
-            symbol_tables: SymbolTables::default(),
-        },
-        analysis_paths: AnalysisPathIndex::default(),
-    };
-
-    assert!(state.snapshot().publish_analysis_output(version, output));
-    assert_eq!(*state.published_analysis_version.borrow(), version);
-    assert!(state.analysis_commit.lock().deferred_source_file_events.is_empty());
-}
-
-#[test]
-fn clearing_analysis_cache_clears_published_path_index() {
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    {
-        let mut commit = state.analysis_commit.lock();
-        commit.discovery_pending = true;
-        commit.analysis_paths = AnalysisPathIndex {
-            resolved_dependencies: FxHashSet::from_iter([PathBuf::from("Dependency.sol")]),
-            existing_unresolved_candidates: FxHashSet::from_iter([PathBuf::from("Unreadable.sol")]),
-            missing_candidates: FxHashSet::from_iter([PathBuf::from("Missing.sol")]),
-        };
-        commit
-            .deferred_source_file_events
-            .insert(PathBuf::from("Pending.sol"), FileChangeType::CHANGED);
-    }
-
-    state.clear_analysis_cache();
-
-    let commit = state.analysis_commit.lock();
-    assert!(commit.cache_invalidated);
-    assert!(!commit.discovery_pending);
-    assert!(commit.analysis_paths.resolved_dependencies.is_empty());
-    assert!(commit.analysis_paths.existing_unresolved_candidates.is_empty());
-    assert!(commit.analysis_paths.missing_candidates.is_empty());
-    assert!(commit.deferred_source_file_events.is_empty());
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn clearing_analysis_cache_rejects_stale_deferred_event_replay() {
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.mark_analysis_pending_for_test();
-    let stale_version = state.analysis_version.load(Ordering::Acquire);
-    state.clear_analysis_cache();
-    let cleared_version = state.analysis_version.load(Ordering::Acquire);
-
-    assert!(matches!(
-        state.on_deferred_source_file_events_ready(DeferredSourceFileEventsReady {
-            version: stale_version,
-            events: vec![(PathBuf::from("Dependency.sol"), FileChangeType::CHANGED)],
-        }),
-        ControlFlow::Continue(())
-    ));
-
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), cleared_version);
-    assert_eq!(*state.published_analysis_version.borrow(), cleared_version);
-    let commit = state.analysis_commit.lock();
-    assert!(commit.cache_invalidated);
-    assert!(!commit.discovery_pending);
-    assert!(commit.deferred_source_file_events.is_empty());
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn unknown_dependency_event_after_cache_clear_starts_recovery() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        import "./generated/Dependency.sol";
-        contract Main is Dependency {}
-
-        //- /generated/Dependency.sol
-        contract Dependency {}
-        "#,
-    );
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config_with_indexing_excludes(&project, &["generated/**"]));
-    state.clear_analysis_cache();
-    let cleared_version = state.analysis_version.load(Ordering::Acquire);
-    project.write_file("/generated/Dependency.sol", "contract Dependency {} contract Latest {}");
-    let path = project.path("/generated/Dependency.sol");
-
-    assert!(matches!(
-        crate::handlers::did_change_watched_files(
-            &mut state,
-            DidChangeWatchedFilesParams {
-                changes: vec![FileEvent {
-                    uri: Url::from_file_path(path).unwrap(),
-                    typ: FileChangeType::CHANGED,
-                }],
-            },
-        ),
-        ControlFlow::Continue(())
-    ));
-
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), cleared_version + 1);
-    let tables = tokio::time::timeout(ASYNC_TEST_TIMEOUT, state.latest_analysis())
-        .await
-        .expect("cache recovery analysis should finish")
-        .unwrap();
-    assert!(!state.analysis_cache_invalidated());
-    assert!(tables.read().workspace_symbols("Latest").iter().any(|symbol| symbol.name == "Latest"));
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn discovery_cleanup_does_not_remove_analysis_handles_for_the_same_epoch() {
-    let version = 1;
-    let discovery = AnalysisTaskKey { version, stage: AnalysisTaskStage::Discovery };
-    let analysis = AnalysisTaskKey { version, stage: AnalysisTaskStage::Analysis };
-    let coordinator = tokio::spawn(std::future::pending::<()>());
-    let worker = tokio::spawn(std::future::pending::<()>());
-    let mut tasks = AnalysisTasks {
-        coordinator: Some((analysis, coordinator.abort_handle())),
-        worker: Some((analysis, worker.abort_handle())),
-        cancellation: None,
-    };
-
-    tasks.clear_worker(discovery);
-    tasks.clear_coordinator(discovery);
-
-    assert!(tasks.worker.as_ref().is_some_and(|(key, _)| *key == analysis));
-    assert!(tasks.coordinator.as_ref().is_some_and(|(key, _)| *key == analysis));
-    tasks.cancel();
-    let _ = coordinator.await;
-    let _ = worker.await;
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn workspace_discovery_router_rejects_stale_and_cancelled_ready_events() {
-    struct DiscoveryStateProbe(oneshot::Sender<(Vec<PathBuf>, bool)>);
-
-    let project = TestProject::from_fixture(
-        r#"
-        //- /stale/Stale.sol
-        contract Stale {}
-
-        //- /latest/Latest.sol
-        contract Latest {}
-        "#,
-    );
-    let discovery_result = |root| {
-        let (_, config) = negotiate_capabilities(project.initialize_params_with_roots(&[root]));
-        config.discover_workspaces(&IndexingCancellation::default()).unwrap()
-    };
-    let stale_result = discovery_result("/stale");
-    let cancelled_result = discovery_result("/stale");
-    let latest_result = discovery_result("/latest");
-    let (_, latest_config) =
-        negotiate_capabilities(project.initialize_params_with_roots(&["/latest"]));
-
-    let (setup_tx, setup_rx) = std_mpsc::sync_channel(1);
-    let (server_main, internal_client) = async_lsp::MainLoop::new_server(move |client| {
-        let mut state = GlobalState::new(client);
-        state.config = Arc::new(latest_config);
-        let (stale_version, stale_progress) = state
-            .begin_analysis(
-                AnalysisMode::Rediscover,
-                Vec::new(),
-                Vec::new(),
-                AnalysisTrigger::External,
-            )
-            .unwrap();
-        let (latest_version, latest_progress) = state
-            .begin_analysis(
-                AnalysisMode::Rediscover,
-                Vec::new(),
-                Vec::new(),
-                AnalysisTrigger::External,
-            )
-            .unwrap();
-        setup_tx
-            .send((
-                stale_version,
-                stale_progress,
-                latest_version,
-                latest_progress,
-                state.published_analysis_version.subscribe(),
-                state.symbol_tables.clone(),
-            ))
-            .unwrap();
-
-        let mut router = crate::new_router_with_state(state);
-        router.event::<DiscoveryStateProbe>(|state, probe| {
-            let roots = state
-                .config
-                .workspaces()
-                .iter()
-                .filter_map(|workspace| workspace.compile_opts().base_path.clone())
-                .collect();
-            let pending = state.analysis_commit.lock().discovery_pending;
-            probe.0.send((roots, pending)).unwrap();
-            ControlFlow::Continue(())
-        });
-        router
-    });
-    let (client_main, server) = async_lsp::MainLoop::new_client(|_| {
-        let mut router = Router::new(());
-        router.notification::<notification::PublishDiagnostics>(|_, _| ControlFlow::Continue(()));
-        router.notification::<notification::LogMessage>(|_, _| ControlFlow::Continue(()));
-        router
-    });
-    let (stale_version, stale_progress, latest_version, latest_progress, mut published, tables) =
-        setup_rx.recv().unwrap();
-
-    let (server_stream, client_stream) = tokio::io::duplex(64 << 10);
-    let (server_rx, server_tx) = tokio::io::split(server_stream);
-    let server_task =
-        tokio::spawn(server_main.run_buffered(server_rx.compat(), server_tx.compat_write()));
-    let (client_rx, client_tx) = tokio::io::split(client_stream);
-    let client_task =
-        tokio::spawn(client_main.run_buffered(client_rx.compat(), client_tx.compat_write()));
-
-    internal_client
-        .emit(WorkspaceDiscoveryReady {
-            version: stale_version,
-            result: stale_result,
-            disk_paths: Vec::new(),
-            progress: stale_progress,
-            cancellation: IndexingCancellation::default(),
-        })
-        .unwrap();
-    let (probe_tx, probe_rx) = oneshot::channel();
-    internal_client.emit(DiscoveryStateProbe(probe_tx)).unwrap();
-    let (roots, pending) = probe_rx.await.unwrap();
-    assert!(roots.is_empty());
-    assert!(pending);
-
-    let cancelled = IndexingCancellation::default();
-    cancelled.cancel();
-    internal_client
-        .emit(WorkspaceDiscoveryReady {
-            version: latest_version,
-            result: cancelled_result,
-            disk_paths: Vec::new(),
-            progress: latest_progress.clone(),
-            cancellation: cancelled,
-        })
-        .unwrap();
-    let (probe_tx, probe_rx) = oneshot::channel();
-    internal_client.emit(DiscoveryStateProbe(probe_tx)).unwrap();
-    let (roots, pending) = probe_rx.await.unwrap();
-    assert!(roots.is_empty());
-    assert!(pending);
-
-    internal_client
-        .emit(WorkspaceDiscoveryReady {
-            version: latest_version,
-            result: latest_result,
-            disk_paths: Vec::new(),
-            progress: latest_progress,
-            cancellation: IndexingCancellation::default(),
-        })
-        .unwrap();
-    tokio::time::timeout(ASYNC_TEST_TIMEOUT, async {
-        while *published.borrow() != latest_version {
-            published.changed().await.unwrap();
-        }
-    })
-    .await
-    .expect("latest workspace discovery should publish analysis");
-
-    assert!(tables.read().workspace_symbols("Stale").is_empty());
-    assert!(tables.read().workspace_symbols("Latest").iter().any(|symbol| symbol.name == "Latest"));
-
-    server.request::<request::Shutdown>(()).await.unwrap();
-    server.notify::<notification::Exit>(()).unwrap();
-    assert!(server_task.await.unwrap().is_ok());
-    assert!(matches!(client_task.await.unwrap(), Err(async_lsp::Error::Eof)));
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn deferred_dependency_change_router_publishes_replacement_analysis() {
-    struct PublishAnalysis {
-        version: usize,
-        output: AnalysisOutput,
-    }
-
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        import "./generated/Dependency.sol";
-        contract Main is Dependency {}
-
-        //- /generated/Dependency.sol
-        contract Dependency {}
-        "#,
-    );
-    let config = config_with_indexing_excludes(&project, &["generated/**"]);
-    let mut batches =
-        snapshot_with_config(config.clone(), project.vfs()).analysis_batches(Vec::new());
-    let old_output =
-        analyze_cancellable(batches.pop().unwrap(), &IndexingCancellation::default()).unwrap();
-    let dependency = project.path("/generated/Dependency.sol");
-    project.write_file("/generated/Dependency.sol", "contract Dependency {} contract Latest {}");
-
-    let (setup_tx, setup_rx) = std_mpsc::sync_channel(1);
-    let (server_main, internal_client) = async_lsp::MainLoop::new_server(move |client| {
-        let mut state = GlobalState::new(client);
-        state.config = Arc::new(config);
-        state.mark_analysis_pending_for_test();
-        let version = state.analysis_version.load(Ordering::Acquire);
-        assert_eq!(
-            state.classify_source_file_event(&dependency, FileChangeType::CHANGED),
-            SourceFileEventDisposition::Deferred
-        );
-        setup_tx
-            .send((
-                version,
-                state.published_analysis_version.subscribe(),
-                state.symbol_tables.clone(),
-            ))
-            .unwrap();
-
-        let mut router = crate::new_router_with_state(state);
-        router.event::<PublishAnalysis>(|state, event| {
-            assert!(!state.snapshot().publish_analysis_output(event.version, event.output));
-            ControlFlow::Continue(())
-        });
-        router
-    });
-    let (client_main, server) = async_lsp::MainLoop::new_client(|_| {
-        let mut router = Router::new(());
-        router.notification::<notification::PublishDiagnostics>(|_, _| ControlFlow::Continue(()));
-        router.notification::<notification::LogMessage>(|_, _| ControlFlow::Continue(()));
-        router
-    });
-    let (version, mut published, tables) = setup_rx.recv().unwrap();
-
-    let (server_stream, client_stream) = tokio::io::duplex(64 << 10);
-    let (server_rx, server_tx) = tokio::io::split(server_stream);
-    let server_task =
-        tokio::spawn(server_main.run_buffered(server_rx.compat(), server_tx.compat_write()));
-    let (client_rx, client_tx) = tokio::io::split(client_stream);
-    let client_task =
-        tokio::spawn(client_main.run_buffered(client_rx.compat(), client_tx.compat_write()));
-
-    internal_client.emit(PublishAnalysis { version, output: old_output }).unwrap();
-    tokio::time::timeout(ASYNC_TEST_TIMEOUT, async {
-        while *published.borrow() <= version {
-            published.changed().await.unwrap();
-        }
-    })
-    .await
-    .expect("replacement dependency analysis should publish");
-
-    assert_eq!(*published.borrow(), version + 1);
-    assert!(tables.read().workspace_symbols("Latest").iter().any(|symbol| symbol.name == "Latest"));
-
-    server.request::<request::Shutdown>(()).await.unwrap();
-    server.notify::<notification::Exit>(()).unwrap();
-    assert!(server_task.await.unwrap().is_ok());
-    assert!(matches!(client_task.await.unwrap(), Err(async_lsp::Error::Eof)));
 }
 
 #[tokio::test(flavor = "current_thread")]
@@ -1856,39 +1104,6 @@ async fn failed_current_analysis_ends_visible_progress() {
 }
 
 #[tokio::test(flavor = "current_thread")]
-async fn failed_current_workspace_discovery_terminates_the_analysis_epoch() {
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    let (version, progress) = state
-        .begin_analysis(AnalysisMode::Rediscover, Vec::new(), Vec::new(), AnalysisTrigger::External)
-        .unwrap();
-    assert!(state.analysis_commit.lock().discovery_pending);
-    let monitor = WorkspaceDiscoveryMonitor {
-        version,
-        disk_paths: Vec::new(),
-        progress,
-        cancellation: IndexingCancellation::default(),
-        analysis_version: state.analysis_version.clone(),
-        published_analysis_version: state.published_analysis_version.clone(),
-        analysis_commit: state.analysis_commit.clone(),
-        client: state.client.clone(),
-        config: state.config.clone(),
-    };
-    let worker = tokio::task::spawn_blocking(|| -> Option<WorkspaceDiscoveryResult> {
-        panic!("test workspace discovery failure")
-    });
-
-    monitor.finish(worker).await;
-
-    tokio::time::timeout(ASYNC_TEST_TIMEOUT, state.latest_analysis())
-        .await
-        .expect("failed discovery should publish its terminal version")
-        .unwrap();
-    let commit = state.analysis_commit.lock();
-    assert!(commit.cache_invalidated);
-    assert!(!commit.discovery_pending);
-}
-
-#[tokio::test(flavor = "current_thread")]
 async fn failed_current_analysis_recovers_after_save() {
     let project = TestProject::from_fixture(
         r#"
@@ -2172,676 +1387,6 @@ fn flycheck_owner(workspace: impl Into<PathBuf>) -> DiagnosticOwner {
 }
 
 #[test]
-fn watched_file_registration_has_global_fallback_patterns() {
-    let [registration] =
-        watched_file_registration_params(&Config::default()).registrations.try_into().unwrap();
-    assert_eq!(registration.id, "solar-watched-files");
-    assert_eq!(registration.method, lsp_types::notification::DidChangeWatchedFiles::METHOD);
-
-    assert_eq!(
-        registration.register_options,
-        Some(serde_json::json!({
-            "watchers": [
-                { "globPattern": "**/*.sol", "kind": WatchKind::Create | WatchKind::Change | WatchKind::Delete },
-                { "globPattern": "**/foundry.toml", "kind": WatchKind::Create | WatchKind::Change | WatchKind::Delete },
-                { "globPattern": "**/remappings.txt", "kind": WatchKind::Create | WatchKind::Change | WatchKind::Delete },
-            ],
-        }))
-    );
-}
-
-#[test]
-fn relative_watched_file_registration_uses_bounded_roots() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /workspace/foundry.toml
-        [profile.default]
-        src = "contracts"
-
-        //- /workspace/contracts/Main.sol
-        contract Main {}
-
-        //- /workspace/node_modules/Dependency.sol
-        contract Dependency {}
-
-        //- /workspace/out/Generated.sol
-        contract Generated {}
-
-        //- /workspace/.hidden/Hidden.sol
-        contract Hidden {}
-
-        //- /workspace/nested/.git/HEAD
-
-        //- /workspace/nested/Nested.sol
-        contract Nested {}
-        "#,
-    );
-    let workspace_root = project.path("/workspace");
-    let source_root = project.path("/workspace/contracts");
-    let mut params = project.initialize_params_with_roots(&["/workspace"]);
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, mut config) = negotiate_capabilities(params);
-
-    let initial = watched_file_registration_params(&config);
-    assert!(watched_file_registration_has_spec(&initial, &workspace_root, "foundry.toml"));
-    assert!(watched_file_registration_has_spec(&initial, &workspace_root, "remappings.txt"));
-    for pattern in ["**/*.sol", "**/foundry.toml", "**/remappings.txt"] {
-        assert!(!watched_file_registration_has_spec(&initial, &workspace_root, pattern));
-    }
-
-    config.rediscover_workspaces();
-    let discovered = watched_file_registration_params(&config);
-    assert!(watched_file_registration_has_spec(&discovered, &source_root, "**/*.sol"));
-    assert!(watched_file_registration_has_spec(&discovered, &source_root, "**/foundry.toml"));
-    assert!(watched_file_registration_has_spec(&discovered, &workspace_root, "foundry.toml"));
-    assert!(watched_file_registration_has_spec(&discovered, &workspace_root, "remappings.txt"));
-    for pattern in ["**/*.sol", "**/foundry.toml", "**/remappings.txt"] {
-        assert!(!watched_file_registration_has_spec(&discovered, &workspace_root, pattern));
-    }
-}
-
-#[test]
-fn relative_watched_file_registration_partitions_root_sources() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry/foundry.toml
-        [profile.default]
-        src = "."
-
-        //- /foundry/Root.sol
-        contract Root {}
-
-        //- /foundry/contracts/Main.sol
-        contract Main {}
-
-        //- /foundry/contracts/core/Core.sol
-        contract Core {}
-
-        //- /foundry/contracts/node_modules/Dependency.sol
-        contract Dependency {}
-
-        //- /foundry/contracts/out/Generated.sol
-        contract Generated {}
-
-        //- /foundry/contracts/.hidden/Hidden.sol
-        contract Hidden {}
-
-        //- /foundry/contracts/vendor/.git/HEAD
-
-        //- /foundry/contracts/vendor/Nested.sol
-        contract Nested {}
-
-        //- /foundry/lib/Dependency.sol
-        contract Dependency {}
-
-        //- /foundry/out/Generated.sol
-        contract Generated {}
-
-        //- /foundry/.hidden/Hidden.sol
-        contract Hidden {}
-
-        //- /foundry/nested/.git/HEAD
-
-        //- /foundry/nested/Nested.sol
-        contract Nested {}
-
-        //- /naked/Root.sol
-        contract Root {}
-
-        //- /naked/contracts/Main.sol
-        contract Main {}
-
-        //- /naked/node_modules/Dependency.sol
-        contract Dependency {}
-        "#,
-    );
-    let foundry_root = project.path("/foundry");
-    let naked_root = project.path("/naked");
-    let mut params = project.initialize_params_with_roots(&["/foundry", "/naked"]);
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, mut config) = negotiate_capabilities(params);
-    config.rediscover_workspaces();
-
-    let registration = watched_file_registration_params(&config);
-    for root in [&foundry_root, &naked_root] {
-        assert!(watched_file_registration_has_spec(&registration, root, "*"));
-        assert!(watched_file_registration_has_spec(&registration, root, "*.sol"));
-        assert!(!watched_file_registration_has_spec(&registration, root, "**/*.sol"));
-        assert_eq!(watched_file_registration_spec_kind(&registration, root, "*"), Some(5));
-        assert_eq!(watched_file_registration_spec_kind(&registration, root, "*.sol"), Some(2));
-    }
-    assert!(watched_file_registration_has_spec(
-        &registration,
-        &foundry_root.join("contracts"),
-        "*"
-    ));
-    assert!(watched_file_registration_has_spec(
-        &registration,
-        &foundry_root.join("contracts"),
-        "*.sol"
-    ));
-    assert!(watched_file_registration_has_spec(
-        &registration,
-        &foundry_root.join("contracts"),
-        "foundry.toml"
-    ));
-    assert!(watched_file_registration_has_spec(
-        &registration,
-        &foundry_root.join("contracts/core"),
-        "**/*.sol"
-    ));
-    assert!(watched_file_registration_has_spec(
-        &registration,
-        &foundry_root.join("contracts/core"),
-        "**/foundry.toml"
-    ));
-    assert!(watched_file_registration_has_spec(
-        &registration,
-        &naked_root.join("contracts"),
-        "**/*.sol"
-    ));
-    for excluded in [
-        foundry_root.join("lib"),
-        foundry_root.join("out"),
-        foundry_root.join(".hidden"),
-        foundry_root.join("nested"),
-        foundry_root.join("contracts/node_modules"),
-        foundry_root.join("contracts/out"),
-        foundry_root.join("contracts/.hidden"),
-        foundry_root.join("contracts/vendor"),
-        naked_root.join("node_modules"),
-    ] {
-        assert!(!watched_file_registration_has_recursive_spec_covering(&registration, &excluded));
-    }
-}
-
-#[test]
-fn relative_watched_file_registration_respects_nested_workspace_ownership() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "."
-
-        //- /nested/foundry.toml
-        [profile.default]
-        src = "src"
-        libs = ["src/vendor"]
-
-        //- /nested/Outside.sol
-        contract Outside {}
-
-        //- /nested/src/Included.sol
-        contract Included {}
-
-        //- /nested/src/generated/Excluded.sol
-        contract Excluded {}
-
-        //- /nested/src/vendor/Dependency.sol
-        contract Dependency {}
-        "#,
-    );
-    let nested_root = project.path("/nested");
-    let nested_source_root = project.path("/nested/src");
-    let mut params = project.initialize_params();
-    params.initialization_options = Some(serde_json::json!({
-        "indexing": { "exclude": ["src/generated/**"] }
-    }));
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, mut config) = negotiate_capabilities(params);
-    config.rediscover_workspaces();
-
-    let registration = watched_file_registration_params(&config);
-    assert!(watched_file_registration_has_spec(&registration, &nested_source_root, "*"));
-    assert!(watched_file_registration_has_spec(&registration, &nested_source_root, "*.sol"));
-    assert!(watched_file_registration_has_spec(&registration, &nested_source_root, "foundry.toml"));
-    assert!(!watched_file_registration_has_spec(&registration, &nested_root, "*.sol"));
-    for excluded in [
-        project.path("/nested/Outside.sol"),
-        project.path("/nested/src/generated"),
-        project.path("/nested/src/vendor"),
-    ] {
-        assert!(!watched_file_registration_has_recursive_spec_covering(&registration, &excluded));
-    }
-}
-
-#[test]
-fn relative_watched_file_registration_omits_excluded_source_root() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "contracts"
-
-        //- /contracts/Main.sol
-        contract Main {}
-        "#,
-    );
-    let source_root = project.path("/contracts");
-    let mut params = project.initialize_params();
-    params.initialization_options = Some(serde_json::json!({
-        "indexing": { "exclude": ["contracts/**"] }
-    }));
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, mut config) = negotiate_capabilities(params);
-    config.rediscover_workspaces();
-
-    let registration = watched_file_registration_params(&config);
-    for pattern in ["*.sol", "**/*.sol"] {
-        assert!(!watched_file_registration_has_spec(&registration, &source_root, pattern));
-    }
-    assert!(!watched_file_registration_has_recursive_spec_covering(&registration, &source_root));
-}
-
-#[test]
-fn watched_file_registration_includes_parent_config_and_external_source_specs() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /repo/foundry.toml
-        [profile.default]
-        src = "../shared"
-
-        //- /repo/workspace/.keep
-        "#,
-    );
-    let mut params = project.initialize_params_with_roots(&["/repo/workspace"]);
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, mut config) = negotiate_capabilities(params);
-    config.rediscover_workspaces();
-    let [registration] =
-        watched_file_registration_params(&config).registrations.try_into().unwrap();
-    let options = registration.register_options.unwrap();
-    let watchers = options["watchers"].as_array().unwrap();
-    let parent_uri = Url::from_file_path(project.path("/repo")).unwrap().to_string();
-    let external_uri = Url::from_file_path(project.path("/shared")).unwrap().to_string();
-
-    assert!(watchers.iter().any(|watcher| {
-        watcher["globPattern"]["baseUri"] == parent_uri
-            && watcher["globPattern"]["pattern"] == "foundry.toml"
-    }));
-    assert!(watchers.iter().any(|watcher| {
-        watcher["globPattern"]["baseUri"] == external_uri
-            && watcher["globPattern"]["pattern"] == "**/*.sol"
-    }));
-}
-
-#[test]
-fn watched_file_specs_add_external_analysis_path_parents_once() {
-    let project = TestProject::new();
-    std::fs::create_dir(project.path("/workspace")).unwrap();
-    let config = project.config_with_roots(&["/workspace"]);
-    let dependency_parent = project.path("/outside");
-    let other_parent = project.path("/other");
-    let analysis_paths = AnalysisPathIndex {
-        resolved_dependencies: FxHashSet::from_iter([dependency_parent.join("Dependency.sol")]),
-        existing_unresolved_candidates: FxHashSet::from_iter([
-            other_parent.join("Candidate.sol"),
-            dependency_parent.join("nested/Deep.sol"),
-        ]),
-        missing_candidates: FxHashSet::from_iter([
-            dependency_parent.join("Missing.sol"),
-            project.path("/workspace/Missing.sol"),
-        ]),
-    };
-
-    let specs = watched_file_specs(&config, &analysis_paths);
-
-    assert_eq!(
-        specs
-            .iter()
-            .filter(|spec| spec.base == dependency_parent && spec.pattern == "*.sol")
-            .count(),
-        1
-    );
-    assert!(specs.iter().any(|spec| spec.base == other_parent && spec.pattern == "*.sol"));
-    assert!(
-        specs
-            .iter()
-            .any(|spec| spec.base == dependency_parent.join("nested") && spec.pattern == "*.sol")
-    );
-    assert!(
-        specs.iter().any(|spec| spec.base == project.path("/workspace") && spec.pattern == "*.sol")
-    );
-    assert!(
-        !specs
-            .iter()
-            .any(|spec| spec.base == project.path("/workspace") && spec.pattern == "**/*.sol")
-    );
-}
-
-#[test]
-fn concurrent_watched_file_updates_keep_desired_specs_and_generation_in_sync() {
-    let project = TestProject::new();
-    std::fs::create_dir(project.path("/workspace")).unwrap();
-    let mut params = project.initialize_params_with_roots(&["/workspace"]);
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, config) = negotiate_capabilities(params);
-    let config = Arc::new(config);
-    let coordinator = Arc::new(WatchedFileRegistrationCoordinator::default());
-    let barrier = Arc::new(Barrier::new(3));
-    let specs = [
-        vec![WatchedFileSpec::new(project.path("/first"), "**/*.sol")],
-        vec![WatchedFileSpec::new(project.path("/second"), "**/*.sol")],
-    ];
-
-    let updates = std::thread::scope(|scope| {
-        let handles = specs.map(|specs| {
-            let config = config.clone();
-            let coordinator = coordinator.clone();
-            let barrier = barrier.clone();
-            scope.spawn(move || {
-                barrier.wait();
-                prepare_watched_file_registration_update(&config, &coordinator, specs, true)
-                    .unwrap()
-            })
-        });
-        barrier.wait();
-        handles.map(|handle| handle.join().unwrap())
-    });
-
-    let desired_specs = coordinator.desired_specs.lock().clone().unwrap();
-    let generation = coordinator.generation.load(Ordering::Acquire);
-    let current = updates.iter().find(|update| update.desired_specs == desired_specs).unwrap();
-    assert_eq!(current.generation, generation);
-}
-
-#[test]
-fn global_fallback_watched_file_update_ignores_spec_changes() {
-    let project = TestProject::new();
-    let mut params = project.initialize_params();
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(false),
-        }),
-        ..Default::default()
-    });
-    let (_, config) = negotiate_capabilities(params);
-    let coordinator = WatchedFileRegistrationCoordinator::default();
-    let first_specs = vec![WatchedFileSpec::new(project.path("/first"), "**/*.sol")];
-    let first = prepare_watched_file_registration_update(&config, &coordinator, first_specs, false)
-        .unwrap();
-
-    assert!(first.desired_specs.is_empty());
-    let generation = first.generation;
-    let second_specs = vec![WatchedFileSpec::new(project.path("/second"), "**/*.sol")];
-    assert!(
-        prepare_watched_file_registration_update(&config, &coordinator, second_specs, true)
-            .is_none()
-    );
-    assert_eq!(coordinator.generation.load(Ordering::Acquire), generation);
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn failed_watched_file_registration_allows_the_same_specs_to_retry() {
-    let project = TestProject::new();
-    std::fs::create_dir(project.path("/workspace")).unwrap();
-    let mut params = project.initialize_params_with_roots(&["/workspace"]);
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, config) = negotiate_capabilities(params);
-    let coordinator = Arc::new(WatchedFileRegistrationCoordinator::default());
-    let specs = config.watched_file_specs();
-    let update =
-        prepare_watched_file_registration_update(&config, &coordinator, specs.clone(), false)
-            .unwrap();
-    let first_generation = update.generation;
-
-    spawn_watched_file_registration_update(&ClientSocket::new_closed(), &coordinator, Some(update));
-    let deadline = Instant::now() + ASYNC_TEST_TIMEOUT;
-    while coordinator.desired_specs.lock().is_some() && Instant::now() < deadline {
-        tokio::task::yield_now().await;
-    }
-    assert!(coordinator.desired_specs.lock().is_none());
-
-    let retry =
-        prepare_watched_file_registration_update(&config, &coordinator, specs, false).unwrap();
-    assert!(retry.generation > first_generation);
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn synchronous_discovery_refreshes_watched_file_specs_before_analysis() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "contracts"
-
-        //- /contracts/Main.sol
-        contract Main {}
-        "#,
-    );
-    let mut params = project.initialize_params();
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, config) = negotiate_capabilities(params);
-    let initial_specs = config.watched_file_specs();
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    *state.watched_file_registration.desired_specs.lock() = Some(initial_specs);
-
-    state.rediscover_workspaces();
-
-    let specs = state.watched_file_registration.desired_specs.lock();
-    assert!(
-        specs
-            .as_ref()
-            .unwrap()
-            .iter()
-            .any(|spec| { spec.base == project.path("/contracts") && spec.pattern == "**/*.sol" })
-    );
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn discovery_and_analysis_refresh_bounded_watched_file_specs() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /repo/foundry.toml
-        [profile.default]
-        src = "../shared"
-
-        //- /repo/workspace/.keep
-        "#,
-    );
-    let mut params = project.initialize_params_with_roots(&["/repo/workspace"]);
-    params.capabilities.workspace = Some(WorkspaceClientCapabilities {
-        did_change_watched_files: Some(DidChangeWatchedFilesClientCapabilities {
-            dynamic_registration: Some(true),
-            relative_pattern_support: Some(true),
-        }),
-        ..Default::default()
-    });
-    let (_, config) = negotiate_capabilities(params);
-    let discovery = config.discover_workspaces(&IndexingCancellation::default()).unwrap();
-    let (server_main, client_socket) = async_lsp::MainLoop::new_server(|_| {
-        let mut router = Router::new(());
-        router.notification::<notification::Exit>(|_, ()| ControlFlow::Break(Ok(())));
-        router
-    });
-    let (events_tx, mut events_rx) = mpsc::unbounded_channel();
-    let (client_main, server_socket) = async_lsp::MainLoop::new_client(move |_| {
-        let mut router = Router::new(events_tx);
-        router.request::<request::RegisterCapability, _>(|events, params| {
-            events.send(WatchedFileClientEvent::Register(params)).unwrap();
-            async { Ok(()) }
-        });
-        router.request::<request::UnregisterCapability, _>(|events, params| {
-            events.send(WatchedFileClientEvent::Unregister(params)).unwrap();
-            async { Ok(()) }
-        });
-        router
-    });
-    let (server_stream, client_stream) = tokio::io::duplex(64 << 10);
-    let (server_rx, server_tx) = tokio::io::split(server_stream);
-    let server_task =
-        tokio::spawn(server_main.run_buffered(server_rx.compat(), server_tx.compat_write()));
-    let (client_rx, client_tx) = tokio::io::split(client_stream);
-    let client_task =
-        tokio::spawn(client_main.run_buffered(client_rx.compat(), client_tx.compat_write()));
-
-    let mut state = GlobalState::new(client_socket);
-    state.config = Arc::new(config);
-    let (version, progress) = state
-        .begin_analysis(AnalysisMode::Rediscover, Vec::new(), Vec::new(), AnalysisTrigger::External)
-        .unwrap();
-
-    assert!(matches!(
-        state.on_workspace_discovery_ready(WorkspaceDiscoveryReady {
-            version,
-            result: discovery,
-            disk_paths: Vec::new(),
-            progress,
-            cancellation: IndexingCancellation::default(),
-        }),
-        ControlFlow::Continue(())
-    ));
-    let discovered_specs = state.watched_file_registration.desired_specs.lock().clone().unwrap();
-    assert!(
-        discovered_specs
-            .iter()
-            .any(|spec| { spec.base == project.path("/repo") && spec.pattern == "foundry.toml" })
-    );
-    assert!(
-        discovered_specs
-            .iter()
-            .any(|spec| { spec.base == project.path("/shared") && spec.pattern == "**/*.sol" })
-    );
-    assert!(matches!(
-        next_watched_file_client_event(&mut events_rx).await,
-        WatchedFileClientEvent::Unregister(params)
-            if params.unregisterations[0].id == "solar-watched-files"
-    ));
-    let WatchedFileClientEvent::Register(discovered_registration) =
-        next_watched_file_client_event(&mut events_rx).await
-    else {
-        panic!("expected discovered watched-file registration")
-    };
-    assert!(watched_file_registration_has_spec(
-        &discovered_registration,
-        &project.path("/repo"),
-        "foundry.toml"
-    ));
-    assert!(watched_file_registration_has_spec(
-        &discovered_registration,
-        &project.path("/shared"),
-        "**/*.sol"
-    ));
-
-    state.analysis_scheduler.tasks.lock().cancel();
-    let dependency_parent = project.path("/outside");
-    let output = AnalysisOutput {
-        result: AnalysisResult {
-            analyzed_documents: AnalyzedDocuments::default(),
-            diagnostics: DiagnosticMap::default(),
-            symbol_tables: SymbolTables::default(),
-        },
-        analysis_paths: AnalysisPathIndex {
-            resolved_dependencies: FxHashSet::from_iter([dependency_parent.join("Dependency.sol")]),
-            ..Default::default()
-        },
-    };
-    assert!(state.snapshot().publish_analysis_output(version, output));
-    let published_specs = state.watched_file_registration.desired_specs.lock().clone().unwrap();
-    assert!(
-        published_specs
-            .iter()
-            .any(|spec| { spec.base == dependency_parent && spec.pattern == "*.sol" })
-    );
-    assert!(
-        !published_specs
-            .iter()
-            .any(|spec| spec.base == dependency_parent && spec.pattern == "**/*.sol")
-    );
-    assert!(matches!(
-        next_watched_file_client_event(&mut events_rx).await,
-        WatchedFileClientEvent::Unregister(_)
-    ));
-    let WatchedFileClientEvent::Register(published_registration) =
-        next_watched_file_client_event(&mut events_rx).await
-    else {
-        panic!("expected analysis watched-file registration")
-    };
-    assert!(watched_file_registration_has_spec(
-        &published_registration,
-        &dependency_parent,
-        "*.sol"
-    ));
-    assert!(!watched_file_registration_has_spec(
-        &published_registration,
-        &dependency_parent,
-        "**/*.sol"
-    ));
-
-    state.clear_analysis_cache();
-    let cleared_specs = state.watched_file_registration.desired_specs.lock().clone().unwrap();
-    assert!(!cleared_specs.iter().any(|spec| spec.base == dependency_parent));
-    assert!(matches!(
-        next_watched_file_client_event(&mut events_rx).await,
-        WatchedFileClientEvent::Unregister(_)
-    ));
-    let WatchedFileClientEvent::Register(cleared_registration) =
-        next_watched_file_client_event(&mut events_rx).await
-    else {
-        panic!("expected cache-clear watched-file registration")
-    };
-    assert!(!watched_file_registration_has_spec(
-        &cleared_registration,
-        &dependency_parent,
-        "**/*.sol"
-    ));
-
-    server_socket.notify::<notification::Exit>(()).unwrap();
-    assert!(server_task.await.unwrap().is_ok());
-    assert!(matches!(client_task.await.unwrap(), Err(async_lsp::Error::Eof)));
-}
-
-#[test]
 fn did_open_tracks_the_source_until_analysis_publishes() {
     let project = TestProject::from_fixture(
         r#"
@@ -2980,756 +1525,6 @@ fn watched_solidity_change_ignores_open_document() {
     assert!(matches!(result, ControlFlow::Continue(())));
     assert_eq!(state.analysis_version.load(Ordering::Acquire), version);
     assert!(state.analysis_scheduler.tasks.lock().coordinator.is_none());
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_solidity_change_ignores_unrelated_excluded_file() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        contract Main {}
-
-        //- /generated/Unrelated.sol
-        contract Unrelated {}
-        "#,
-    );
-    let path = project.path("/generated/Unrelated.sol");
-    let uri = Url::from_file_path(path).unwrap();
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config_with_indexing_excludes(&project, &["generated/**"]));
-    let version = state.analysis_version.load(Ordering::Acquire);
-
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: vec![FileEvent { uri, typ: FileChangeType::CHANGED }],
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), version);
-    assert!(state.analysis_scheduler.tasks.lock().coordinator.is_none());
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_excluded_manifest_events_do_not_schedule_analysis() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        contract Main {}
-
-        //- /node_modules/package/foundry.toml
-        [profile.default]
-        src = "src"
-        "#,
-    );
-
-    for typ in [FileChangeType::CREATED, FileChangeType::CHANGED, FileChangeType::DELETED] {
-        let mut state = GlobalState::new(ClientSocket::new_closed());
-        state.config = Arc::new(project.config());
-        let version = state.analysis_version.load(Ordering::Acquire);
-        let uri = Url::from_file_path(project.path("/node_modules/package/foundry.toml")).unwrap();
-
-        let result = crate::handlers::did_change_watched_files(
-            &mut state,
-            DidChangeWatchedFilesParams { changes: vec![FileEvent { uri, typ }] },
-        );
-
-        assert!(matches!(result, ControlFlow::Continue(())));
-        assert_eq!(state.analysis_version.load(Ordering::Acquire), version);
-        assert!(state.analysis_scheduler.tasks.lock().coordinator.is_none());
-    }
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_nested_manifest_create_discovers_the_project() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-
-        //- /packages/app/foundry.toml
-
-        //- /packages/app/generated/.keep
-        "#,
-    );
-    let mut params = project.initialize_params();
-    params.initialization_options = Some(serde_json::json!({
-        "indexing": { "exclude": ["packages/app/generated/**"] }
-    }));
-    let (_, mut config) = negotiate_capabilities(params);
-    config.rediscover_workspaces();
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    project.write_file(
-        "/packages/app/generated/foundry.toml",
-        r#"
-        [profile.default]
-        src = "src"
-        "#,
-    );
-    project.write_file("/packages/app/generated/src/Nested.sol", "contract Nested {}");
-    let uri = Url::from_file_path(project.path("/packages/app/generated/foundry.toml")).unwrap();
-
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: vec![FileEvent { uri, typ: FileChangeType::CREATED }],
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    let tables = tokio::time::timeout(ASYNC_TEST_TIMEOUT, state.latest_analysis())
-        .await
-        .expect("nested manifest analysis should finish")
-        .unwrap();
-    assert!(tables.read().workspace_symbols("Nested").iter().any(|symbol| symbol.name == "Nested"));
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn created_directory_under_overlapping_source_root_schedules_analysis() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "lib"
-        "#,
-    );
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(project.config());
-    let version = state.analysis_version.load(Ordering::Acquire);
-    let path = project.path("/lib/generated");
-    std::fs::create_dir_all(&path).unwrap();
-
-    let result = crate::handlers::did_create_files(
-        &mut state,
-        CreateFilesParams {
-            files: vec![FileCreate { uri: Url::from_file_path(path).unwrap().to_string() }],
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    let actual_version = state.analysis_version.load(Ordering::Acquire);
-    state.analysis_scheduler.tasks.lock().cancel();
-    assert_eq!(actual_version, version + 1);
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_created_source_under_overlapping_root_is_tracked() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "lib"
-        "#,
-    );
-    let config = project.config();
-    let path = project.path("/lib/Created.sol");
-    project.write_file("/lib/Created.sol", "contract Created {}");
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: vec![FileEvent {
-                uri: Url::from_file_path(&path).unwrap(),
-                typ: FileChangeType::CREATED,
-            }],
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), 1);
-    assert_eq!(state.config.tracked_source_files_under(&[project.path("/lib")]), [path]);
-    state.analysis_scheduler.tasks.lock().cancel();
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_created_directory_under_partitioned_root_is_rediscovered() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "."
-
-        //- /lib/Dependency.sol
-        contract Dependency {}
-        "#,
-    );
-    let config = project.config();
-    let directory = project.path("/new");
-    let source = project.path("/new/New.sol");
-    let unrelated = project.path("/README.md");
-    project.write_file("/README.md", "notes");
-    project.write_file("/new/New.sol", "contract New {}");
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: vec![FileEvent {
-                uri: Url::from_file_path(unrelated).unwrap(),
-                typ: FileChangeType::CREATED,
-            }],
-        },
-    );
-    assert!(matches!(result, ControlFlow::Continue(())));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), 0);
-
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: vec![FileEvent {
-                uri: Url::from_file_path(directory).unwrap(),
-                typ: FileChangeType::CREATED,
-            }],
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), 1);
-    assert_eq!(state.config.tracked_source_files_under(&[project.root().to_path_buf()]), [source]);
-    state.analysis_scheduler.tasks.lock().cancel();
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_deleted_directory_under_partitioned_root_is_rediscovered() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "."
-
-        //- /lib/Dependency.sol
-        contract Dependency {}
-
-        //- /old/Old.sol
-        contract Old {}
-        "#,
-    );
-    let config = project.config();
-    let directory = project.path("/old");
-    std::fs::remove_dir_all(&directory).unwrap();
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: vec![FileEvent {
-                uri: Url::from_file_path(directory).unwrap(),
-                typ: FileChangeType::DELETED,
-            }],
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), 1);
-    assert!(state.config.tracked_source_files_under(&[project.root().to_path_buf()]).is_empty());
-    state.analysis_scheduler.tasks.lock().cancel();
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn created_directory_below_default_exclude_does_not_schedule_analysis() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        contract Main {}
-        "#,
-    );
-    let path = project.path("/node_modules/package/generated");
-    std::fs::create_dir_all(&path).unwrap();
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(project.config());
-    let version = state.analysis_version.load(Ordering::Acquire);
-
-    let result = crate::handlers::did_create_files(
-        &mut state,
-        CreateFilesParams {
-            files: vec![FileCreate { uri: Url::from_file_path(path).unwrap().to_string() }],
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    let actual_version = state.analysis_version.load(Ordering::Acquire);
-    state.analysis_scheduler.tasks.lock().cancel();
-    assert_eq!(actual_version, version);
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_excluded_dependency_change_and_delete_schedule_analysis() {
-    for typ in [FileChangeType::CHANGED, FileChangeType::DELETED] {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /Main.sol
-            import "./generated/Dependency.sol";
-            contract Main is Dependency {}
-
-            //- /generated/Dependency.sol
-            contract Dependency {}
-            "#,
-        );
-        let config = config_with_indexing_excludes(&project, &["generated/**"]);
-        let mut batches =
-            snapshot_with_config(config.clone(), project.vfs()).analysis_batches(Vec::new());
-        let output =
-            analyze_cancellable(batches.pop().unwrap(), &IndexingCancellation::default()).unwrap();
-        let mut state = GlobalState::new(ClientSocket::new_closed());
-        state.config = Arc::new(config);
-        state.snapshot().publish_analysis_output(0, output);
-        let path = project.path("/generated/Dependency.sol");
-        let uri = Url::from_file_path(path).unwrap();
-
-        let result = crate::handlers::did_change_watched_files(
-            &mut state,
-            DidChangeWatchedFilesParams { changes: vec![FileEvent { uri, typ }] },
-        );
-
-        assert!(matches!(result, ControlFlow::Continue(())));
-        assert_eq!(state.analysis_version.load(Ordering::Acquire), 1);
-        state.analysis_scheduler.tasks.lock().cancel();
-    }
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_external_source_create_change_and_delete_schedule_analysis() {
-    for typ in [FileChangeType::CREATED, FileChangeType::CHANGED, FileChangeType::DELETED] {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /project/foundry.toml
-            [profile.default]
-            src = "../shared"
-            "#,
-        );
-        let path = project.path("/shared/External.sol");
-        if typ != FileChangeType::CREATED {
-            project.write_file("/shared/External.sol", "contract External {}");
-        }
-        let config = project.config_with_roots(&["/project"]);
-        if typ == FileChangeType::CREATED {
-            project.write_file("/shared/External.sol", "contract External {}");
-        } else if typ == FileChangeType::DELETED {
-            std::fs::remove_file(&path).unwrap();
-        }
-        let mut state = GlobalState::new(ClientSocket::new_closed());
-        state.config = Arc::new(config);
-
-        assert!(matches!(
-            crate::handlers::did_change_watched_files(
-                &mut state,
-                DidChangeWatchedFilesParams {
-                    changes: vec![FileEvent { uri: Url::from_file_path(&path).unwrap(), typ }],
-                },
-            ),
-            ControlFlow::Continue(())
-        ));
-
-        assert_eq!(state.analysis_version.load(Ordering::Acquire), 1);
-        let tracked = state.config.tracked_source_files_under(&[project.path("/shared")]);
-        if typ == FileChangeType::DELETED {
-            assert!(tracked.is_empty());
-        } else {
-            assert_eq!(tracked, [path]);
-        }
-        state.analysis_scheduler.tasks.lock().cancel();
-    }
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn unknown_dependency_event_is_deferred_while_analysis_is_pending() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        contract Main {}
-
-        //- /generated/Dependency.sol
-        contract Dependency {}
-        "#,
-    );
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config_with_indexing_excludes(&project, &["generated/**"]));
-    state.mark_analysis_pending_for_test();
-    let path = project.path("/generated/Dependency.sol");
-
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: vec![FileEvent {
-                uri: Url::from_file_path(&path).unwrap(),
-                typ: FileChangeType::CHANGED,
-            }],
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), 1);
-    assert_eq!(
-        state.analysis_commit.lock().deferred_source_file_events.get(&path),
-        Some(&FileChangeType::CHANGED)
-    );
-}
-
-#[test]
-fn did_create_defers_a_candidate_first_learned_by_pending_analysis() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        import "./generated/Dependency.sol";
-        contract Main is Dependency {}
-        "#,
-    );
-    let config = config_with_indexing_excludes(&project, &["generated/**"]);
-    let mut batches =
-        snapshot_with_config(config.clone(), project.vfs()).analysis_batches(Vec::new());
-    let output =
-        analyze_cancellable(batches.pop().unwrap(), &IndexingCancellation::default()).unwrap();
-    let path = project.path("/generated/Dependency.sol");
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    state.mark_analysis_pending_for_test();
-    let version = state.analysis_version.load(Ordering::Acquire);
-    project.write_file("/generated/Dependency.sol", "contract Dependency {}");
-
-    assert!(matches!(
-        crate::handlers::did_create_files(
-            &mut state,
-            CreateFilesParams {
-                files: vec![FileCreate { uri: Url::from_file_path(&path).unwrap().to_string() }],
-            },
-        ),
-        ControlFlow::Continue(())
-    ));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), version);
-    assert_eq!(
-        state.analysis_commit.lock().deferred_source_file_events.get(&path),
-        Some(&FileChangeType::CREATED)
-    );
-    assert!(!state.snapshot().publish_analysis_output(version, output));
-}
-
-#[test]
-fn did_delete_defers_a_dependency_first_learned_by_pending_analysis() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        import "./generated/Dependency.sol";
-        contract Main is Dependency {}
-
-        //- /generated/Dependency.sol
-        contract Dependency {}
-        "#,
-    );
-    let config = config_with_indexing_excludes(&project, &["generated/**"]);
-    let mut batches =
-        snapshot_with_config(config.clone(), project.vfs()).analysis_batches(Vec::new());
-    let output =
-        analyze_cancellable(batches.pop().unwrap(), &IndexingCancellation::default()).unwrap();
-    let path = project.path("/generated/Dependency.sol");
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    state.mark_analysis_pending_for_test();
-    let version = state.analysis_version.load(Ordering::Acquire);
-    std::fs::remove_file(&path).unwrap();
-
-    assert!(matches!(
-        crate::handlers::did_delete_files(
-            &mut state,
-            DeleteFilesParams {
-                files: vec![FileDelete { uri: Url::from_file_path(&path).unwrap().to_string() }],
-            },
-        ),
-        ControlFlow::Continue(())
-    ));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), version);
-    assert_eq!(
-        state.analysis_commit.lock().deferred_source_file_events.get(&path),
-        Some(&FileChangeType::DELETED)
-    );
-    assert!(!state.snapshot().publish_analysis_output(version, output));
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn source_events_during_initial_discovery_are_replayed_after_policy_is_known() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /project/foundry.toml
-        [profile.default]
-        src = "lib"
-        libs = ["vendor"]
-        "#,
-    );
-    let (_, config) = negotiate_capabilities(project.initialize_params_with_roots(&["/project"]));
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    let (version, progress) = state
-        .begin_analysis(AnalysisMode::Rediscover, Vec::new(), Vec::new(), AnalysisTrigger::External)
-        .unwrap();
-    let discovery = state.config.discover_workspaces(&IndexingCancellation::default()).unwrap();
-
-    project.write_file("/project/lib/Active.sol", "contract Active {}");
-    project.write_file("/project/node_modules/Ignored.sol", "contract Ignored {}");
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: [
-                project.path("/project/lib/Active.sol"),
-                project.path("/project/node_modules/Ignored.sol"),
-            ]
-            .into_iter()
-            .map(|path| FileEvent {
-                uri: Url::from_file_path(path).unwrap(),
-                typ: FileChangeType::CREATED,
-            })
-            .collect(),
-        },
-    );
-
-    assert!(matches!(result, ControlFlow::Continue(())));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), version);
-    assert!(matches!(
-        state.on_workspace_discovery_ready(WorkspaceDiscoveryReady {
-            version,
-            result: discovery,
-            disk_paths: Vec::new(),
-            progress,
-            cancellation: IndexingCancellation::default(),
-        }),
-        ControlFlow::Continue(())
-    ));
-
-    let tables = tokio::time::timeout(ASYNC_TEST_TIMEOUT, state.latest_analysis())
-        .await
-        .expect("initial discovery analysis should finish")
-        .unwrap();
-    assert!(tables.read().workspace_symbols("Active").iter().any(|symbol| symbol.name == "Active"));
-    assert!(tables.read().workspace_symbols("Ignored").is_empty());
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn source_events_during_discovery_are_deferred_with_existing_workspaces() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /existing/Existing.sol
-        contract Existing {}
-        "#,
-    );
-    let new_root = project.path("/new");
-    std::fs::create_dir(&new_root).unwrap();
-    let (_, mut config) =
-        negotiate_capabilities(project.initialize_params_with_roots(&["/existing"]));
-    config.rediscover_workspaces();
-    assert!(!config.workspaces().is_empty());
-    config.add_workspaces([new_root.clone()]);
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    let (version, progress) = state
-        .begin_analysis(AnalysisMode::Rediscover, Vec::new(), Vec::new(), AnalysisTrigger::External)
-        .unwrap();
-    let discovery = state.config.discover_workspaces(&IndexingCancellation::default()).unwrap();
-    let path = project.path("/new/Active.sol");
-    project.write_file("/new/Active.sol", "contract Active {}");
-
-    assert!(matches!(
-        crate::handlers::did_change_watched_files(
-            &mut state,
-            DidChangeWatchedFilesParams {
-                changes: vec![
-                    FileEvent {
-                        uri: Url::from_file_path(&path).unwrap(),
-                        typ: FileChangeType::CREATED,
-                    },
-                    FileEvent {
-                        uri: Url::from_file_path(&path).unwrap(),
-                        typ: FileChangeType::CHANGED,
-                    },
-                ],
-            },
-        ),
-        ControlFlow::Continue(())
-    ));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), version);
-    assert_eq!(
-        state.analysis_commit.lock().deferred_source_file_events.get(&path),
-        Some(&FileChangeType::CHANGED)
-    );
-
-    assert!(matches!(
-        state.on_workspace_discovery_ready(WorkspaceDiscoveryReady {
-            version,
-            result: discovery,
-            disk_paths: Vec::new(),
-            progress,
-            cancellation: IndexingCancellation::default(),
-        }),
-        ControlFlow::Continue(())
-    ));
-    assert_eq!(
-        state.config.tracked_source_files_under(std::slice::from_ref(&new_root)),
-        std::slice::from_ref(&path)
-    );
-    let tables = tokio::time::timeout(ASYNC_TEST_TIMEOUT, state.latest_analysis())
-        .await
-        .expect("rediscovered source analysis should finish")
-        .unwrap();
-    assert!(tables.read().workspace_symbols("Active").iter().any(|symbol| symbol.name == "Active"));
-    drop(tables);
-
-    state.recompute_after_source_changes(vec![project.path("/existing/Existing.sol")]);
-    let tables = tokio::time::timeout(ASYNC_TEST_TIMEOUT, state.latest_analysis())
-        .await
-        .expect("subsequent analysis should finish")
-        .unwrap();
-    assert!(tables.read().workspace_symbols("Active").iter().any(|symbol| symbol.name == "Active"));
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_existing_unresolved_candidate_change_and_delete_schedule_analysis() {
-    for typ in [FileChangeType::CHANGED, FileChangeType::DELETED] {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /foundry.toml
-            [profile.default]
-            src = "src"
-            libs = ["lib-one", "lib-two"]
-
-            //- /src/Main.sol
-            import "Dependency.sol";
-            contract Main is Dependency {}
-
-            //- /lib-one/Dependency.sol
-            contract Dependency {}
-
-            //- /lib-two/Dependency.sol
-            contract Dependency {}
-            "#,
-        );
-        let config = project.config();
-        let mut batches =
-            snapshot_with_config(config.clone(), project.vfs()).analysis_batches(Vec::new());
-        let output =
-            analyze_cancellable(batches.pop().unwrap(), &IndexingCancellation::default()).unwrap();
-        let path = project.path("/lib-one/Dependency.sol");
-        match typ {
-            FileChangeType::CHANGED => {
-                std::fs::write(&path, "contract Dependency { uint x; }").unwrap()
-            }
-            FileChangeType::DELETED => std::fs::remove_file(&path).unwrap(),
-            _ => unreachable!(),
-        }
-        let uri = Url::from_file_path(path).unwrap();
-        let mut state = GlobalState::new(ClientSocket::new_closed());
-        state.config = Arc::new(config);
-        state.snapshot().publish_analysis_output(0, output);
-
-        let result = crate::handlers::did_change_watched_files(
-            &mut state,
-            DidChangeWatchedFilesParams { changes: vec![FileEvent { uri, typ }] },
-        );
-
-        assert!(matches!(result, ControlFlow::Continue(())));
-        let actual_version = state.analysis_version.load(Ordering::Acquire);
-        state.analysis_scheduler.tasks.lock().cancel();
-        assert_eq!(actual_version, 1);
-    }
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_missing_excluded_dependency_recovers_only_on_create() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        import "./generated/Missing.sol";
-        contract Main is Missing {}
-        "#,
-    );
-    let config = config_with_indexing_excludes(&project, &["generated/**"]);
-    let mut batches =
-        snapshot_with_config(config.clone(), project.vfs()).analysis_batches(Vec::new());
-    let output =
-        analyze_cancellable(batches.pop().unwrap(), &IndexingCancellation::default()).unwrap();
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    state.snapshot().publish_analysis_output(0, output);
-    let path = project.path("/generated/Missing.sol");
-    let uri = Url::from_file_path(&path).unwrap();
-
-    for typ in [FileChangeType::CHANGED, FileChangeType::DELETED] {
-        let result = crate::handlers::did_change_watched_files(
-            &mut state,
-            DidChangeWatchedFilesParams { changes: vec![FileEvent { uri: uri.clone(), typ }] },
-        );
-        assert!(matches!(result, ControlFlow::Continue(())));
-        assert_eq!(state.analysis_version.load(Ordering::Acquire), 0);
-    }
-
-    project.write_file("/generated/Missing.sol", "contract Missing {}");
-    let result = crate::handlers::did_change_watched_files(
-        &mut state,
-        DidChangeWatchedFilesParams {
-            changes: vec![FileEvent { uri, typ: FileChangeType::CREATED }],
-        },
-    );
-    assert!(matches!(result, ControlFlow::Continue(())));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), 1);
-
-    let tables = tokio::time::timeout(ASYNC_TEST_TIMEOUT, state.latest_analysis())
-        .await
-        .expect("created import candidate should be analyzed")
-        .unwrap();
-    assert!(
-        tables.read().workspace_symbols("Missing").iter().any(|symbol| symbol.name == "Missing")
-    );
-}
-
-#[tokio::test(flavor = "current_thread")]
-async fn watched_missing_candidate_change_supersedes_pending_create_analysis() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /Main.sol
-        import "./generated/Missing.sol";
-        contract Main is Missing {}
-        "#,
-    );
-    let config = config_with_indexing_excludes(&project, &["generated/**"]);
-    let mut batches =
-        snapshot_with_config(config.clone(), project.vfs()).analysis_batches(Vec::new());
-    let output =
-        analyze_cancellable(batches.pop().unwrap(), &IndexingCancellation::default()).unwrap();
-    let mut state = GlobalState::new(ClientSocket::new_closed());
-    state.config = Arc::new(config);
-    state.snapshot().publish_analysis_output(0, output);
-    let path = project.path("/generated/Missing.sol");
-    let uri = Url::from_file_path(&path).unwrap();
-
-    project.write_file("/generated/Missing.sol", "contract Missing {}");
-    assert!(matches!(
-        crate::handlers::did_change_watched_files(
-            &mut state,
-            DidChangeWatchedFilesParams {
-                changes: vec![FileEvent { uri: uri.clone(), typ: FileChangeType::CREATED }],
-            },
-        ),
-        ControlFlow::Continue(())
-    ));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), 1);
-
-    project.write_file("/generated/Missing.sol", "contract Missing { uint latest; }");
-    assert!(matches!(
-        crate::handlers::did_change_watched_files(
-            &mut state,
-            DidChangeWatchedFilesParams {
-                changes: vec![FileEvent { uri, typ: FileChangeType::CHANGED }],
-            },
-        ),
-        ControlFlow::Continue(())
-    ));
-    assert_eq!(state.analysis_version.load(Ordering::Acquire), 2);
-
-    tokio::time::timeout(ASYNC_TEST_TIMEOUT, state.latest_analysis())
-        .await
-        .expect("replacement dependency analysis should finish")
-        .unwrap();
 }
 
 #[test]
@@ -4398,60 +2193,6 @@ fn analysis_batches_use_cached_workspace_source_files() {
 }
 
 #[test]
-fn analysis_batches_index_closed_sources_when_source_root_overlaps_library() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "lib"
-
-        //- /lib/Main.sol
-        contract Main {}
-        "#,
-    );
-    let snapshot = snapshot_with_config(project.config(), Vfs::default());
-
-    let mut batches = snapshot.analysis_batches(Vec::new());
-    let batch = batches.pop().unwrap();
-
-    assert_eq!(batch.files, vec![(project.path("/lib/Main.sol"), "contract Main {}".into())]);
-}
-
-#[test]
-fn analysis_batches_keep_sources_below_import_only_manifest_corridors() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "lib/contracts"
-
-        //- /lib/foundry.toml
-        [profile.default]
-        src = "other"
-
-        //- /lib/contracts/Main.sol
-        contract Main {}
-
-        //- /lib/dependency/Dependency.sol
-        contract Dependency {}
-        "#,
-    );
-    let config = project.config();
-    assert!(!config.workspaces().iter().any(|workspace| {
-        workspace.compile_opts().base_path.as_deref() == Some(project.path("/lib").as_path())
-    }));
-    let snapshot = snapshot_with_config(config, Vfs::default());
-
-    let mut batches = snapshot.analysis_batches(Vec::new());
-    let batch = batches.pop().unwrap();
-
-    assert_eq!(
-        batch.files,
-        vec![(project.path("/lib/contracts/Main.sol"), "contract Main {}".into())]
-    );
-}
-
-#[test]
 fn analysis_batches_assign_open_files_to_most_specific_workspace() {
     let project = TestProject::from_fixture(
         r#"
@@ -4476,65 +2217,6 @@ fn analysis_batches_assign_open_files_to_most_specific_workspace() {
 
     assert!(!outer_batch.files.iter().any(|(path, _)| path == &source_path));
     assert_eq!(inner_batch.files, vec![(source_path, "contract A {}".into())]);
-}
-
-#[test]
-fn workspace_discovery_rechecks_sources_against_the_owning_workspace_policy() {
-    let project = TestProject::from_fixture(
-        r#"
-        //- /foundry.toml
-        [profile.default]
-        src = "."
-
-        //- /nested/foundry.toml
-        [profile.default]
-        src = "src"
-        libs = ["src/vendor"]
-
-        //- /nested/src/Included.sol
-        contract Included {}
-
-        //- /nested/src/generated/Excluded.sol
-        contract Excluded {}
-
-        //- /nested/src/vendor/Dependency.sol
-        contract Dependency {}
-
-        //- /nested/Outside.sol
-        contract Outside {}
-        "#,
-    );
-    let config = config_with_indexing_excludes(&project, &["src/generated/**"]);
-    let nested_root = project.path("/nested");
-    let outer_workspace = config
-        .workspaces()
-        .iter()
-        .find(|workspace| workspace.compile_opts().base_path.as_deref() == Some(project.root()))
-        .unwrap();
-    let nested_workspace = config
-        .workspaces()
-        .iter()
-        .find(|workspace| workspace.compile_opts().base_path.as_deref() == Some(&nested_root))
-        .unwrap();
-    assert!(outer_workspace.source_files().is_empty());
-    assert_eq!(nested_workspace.source_files(), [project.path("/nested/src/Included.sol")]);
-    assert_eq!(config.index_metrics().eager, 1);
-
-    let batches = snapshot_with_config(config, Vfs::default()).analysis_batches(Vec::new());
-    let outer_batch = batches
-        .iter()
-        .find(|batch| batch.opts.base_path.as_deref() == Some(project.root()))
-        .unwrap();
-    let nested_batch = batches
-        .iter()
-        .find(|batch| batch.opts.base_path.as_deref() == Some(nested_root.as_path()))
-        .unwrap();
-
-    assert!(outer_batch.files.iter().all(|(path, _)| !path.starts_with(&nested_root)));
-    assert_eq!(
-        nested_batch.files,
-        vec![(project.path("/nested/src/Included.sol"), "contract Included {}".into())]
-    );
 }
 
 #[test]
